Move telemetry outside of configure
