+++ src/configure/configure.ts
@@ -10,13 +10,13 @@ import { Messages } from './resources/messages';
 import { ServiceConnectionHelper } from './helper/devOps/serviceConnectionHelper';
 import { SourceOptions, RepositoryProvider, extensionVariables, WizardInputs, WebAppKind, PipelineTemplate, QuickPickItemWithData, GitRepositoryParameters, GitBranchDetails, TargetResourceType } from './model/models';
 import { TracePoints } from './resources/tracePoints';
-import { TelemetryKeys } from './resources/telemetryKeys';
+import { TelemetryKeys } from '../helpers/telemetryKeys';
 import * as constants from './resources/constants';
 import * as path from 'path';
 import * as templateHelper from './helper/templateHelper';
 import * as utils from 'util';
 import * as vscode from 'vscode';
-import { telemetryHelper } from './helper/telemetryHelper';
+import { telemetryHelper } from '../helpers/telemetryHelper';
 import { ControlProvider } from './helper/controlProvider';
 import { GitHubProvider } from './helper/gitHubHelper';
 import { getSubscriptionSession } from './helper/azureSessionHelper';
@@ -392,10 +392,10 @@ class PipelineConfigurer {
             let selectedSubscription: QuickPickItemWithData = await this.controlProvider.showQuickPick(constants.SelectSubscription, subscriptionList, { placeHolder: Messages.selectSubscription });
             this.inputs.targetResource.subscriptionId = selectedSubscription.data.subscription.subscriptionId;
             this.inputs.azureSession = getSubscriptionSession(this.inputs.targetResource.subscriptionId);
-            
+
             // show available resources and get the chosen one
             this.appServiceClient = new AppServiceClient(this.inputs.azureSession.credentials, this.inputs.azureSession.tenantId, this.inputs.azureSession.environment.portalUrl, this.inputs.targetResource.subscriptionId);
-            
+
             let resourceArray: Promise<Array<{label: string, data: GenericResource}>> = null;
             let selectAppText: string = "";
             let placeHolderText: string = "";
@@ -415,7 +415,7 @@ class PipelineConfigurer {
                 resourceArray,
                 { placeHolder:  placeHolderText },
                 TelemetryKeys.WebAppListCount);
-    
+
             this.inputs.targetResource.resource = selectedResource.data;
         } else if(subscriptionList.length > 0 ) {
             this.inputs.targetResource.subscriptionId = subscriptionList[0].data.subscription.subscriptionId;
